+++ raspberry-pi/app/motors/__init__.py
@@ -1,6 +1,7 @@
 """Hardware-separated motor interfaces and motion policy."""
 
 from .homing import HomingError, HomingResult, home_stepper, home_stepper_for_seconds
+from .servo import GateConfig, LgpioServo, MockServo, ServoError, ServoGate
 from .stepper import LgpioStepper, StepperError
 
 __all__ = [
@@ -10,4 +11,9 @@ __all__ = [
     "StepperError",
     "home_stepper",
     "home_stepper_for_seconds",
+    "GateConfig",
+    "LgpioServo",
+    "MockServo",
+    "ServoError",
+    "ServoGate",
 ]